Feat: 공지방 생성 퍼블리싱 수정 및 api 연동 중
